[Rollout] Production rollout

+++ src/Microsoft.DotNet.Darc/src/Darc/Operations/GatherDropOperation.cs
@@ -13,6 +13,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Net;
 using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
@@ -416,6 +417,30 @@ namespace Microsoft.DotNet.Darc.Operations
         }
 
         /// <summary>
+        /// Filter any released builds if the user specified --skip-released
+        /// </summary>
+        /// <param name="inputBuilds">Input builds</param>
+        /// <returns>Builds to download</returns>
+        private IEnumerable<Build> FilterReleasedBuilds(IEnumerable<Build> builds)
+        {
+            if (_options.SkipReleased)
+            {
+                var releasedBuilds = builds.Where(build => build.Released);
+                var nonReleasedBuilds = builds.Where(build => !build.Released);
+                foreach (var build in releasedBuilds)
+                {
+                    if (build.Released)
+                    {
+                        Console.WriteLine($"  Skipping download of released build {build.AzureDevOpsBuildNumber} of {build.GitHubRepository ?? build.AzureDevOpsRepository} @ {build.Commit}");
+                    }
+                }
+                return nonReleasedBuilds;
+            }
+
+            return builds;
+        }
+
+        /// <summary>
         ///     Build the list of builds that will need their assets downloaded.
         /// </summary>
         /// <returns>List of builds to download</returns>
@@ -442,7 +467,7 @@ namespace Microsoft.DotNet.Darc.Operations
                 return new InputBuilds()
                 {
                     Successful = true,
-                    Builds = new List<Build>() { rootBuild }
+                    Builds = FilterReleasedBuilds(new List<Build>() { rootBuild })
                 };
             }
 
@@ -499,6 +524,8 @@ namespace Microsoft.DotNet.Darc.Operations
                 builds.Add(build);
             }
 
+            var filteredBuilds = FilterReleasedBuilds(builds);
+
             if (graph.DependenciesMissingBuilds.Any())
             {
                 Console.WriteLine("Dependencies missing builds:");
@@ -511,7 +538,7 @@ namespace Microsoft.DotNet.Darc.Operations
                     return new InputBuilds()
                     {
                         Successful = false,
-                        Builds = builds
+                        Builds = filteredBuilds
                     };
                 }
             }
@@ -519,7 +546,7 @@ namespace Microsoft.DotNet.Darc.Operations
             return new InputBuilds()
             {
                 Successful = true,
-                Builds = builds.ToList()
+                Builds = filteredBuilds
             };
         }
 
@@ -574,6 +601,11 @@ namespace Microsoft.DotNet.Darc.Operations
             using (HttpClient client = new HttpClient(new HttpClientHandler { CheckCertificateRevocationList = true }))
             {
                 var assets = await remote.GetAssetsAsync(buildId: build.Id, nonShipping: (!_options.IncludeNonShipping ? (bool?)false : null));
+                if (!string.IsNullOrEmpty(_options.AssetFilter))
+                {
+                    assets = assets.Where(asset => Regex.IsMatch(asset.Name, _options.AssetFilter));
+                }
+
                 using (var clientThrottle = new SemaphoreSlim(_options.MaxConcurrentDownloads, _options.MaxConcurrentDownloads))
                 {
                     await Task.WhenAll(assets.Select(async asset =>
@@ -881,7 +913,10 @@ namespace Microsoft.DotNet.Darc.Operations
                 string finalUri = assetLocation.Location.Substring(0, assetLocation.Location.Length - "index.json".Length);
                 finalUri += $"flatcontainer/{name}/{version}/{name}.{version}.nupkg";
 
-                if (await DownloadFileAsync(client, finalUri, fullTargetPath, errors, downloadOutput))
+                using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+                var cancellationToken = cancellationTokenSource.Token;
+
+                if (await DownloadFileAsync(client, finalUri, null, fullTargetPath, errors, downloadOutput, cancellationToken))
                 {
                     return new DownloadedAsset()
                     {
@@ -929,9 +964,21 @@ namespace Microsoft.DotNet.Darc.Operations
                                                                                 List<string> errors,
                                                                                 StringBuilder downloadOutput)
         {
-            string packageContentUrl = $"https://pkgs.dev.azure.com/{feedAccount}/{feedVisibility}_apis/packaging/feeds/{feedName}/nuget/packages/{asset.Name}/versions/{asset.Version}/content";
+            string assetName = asset.Name;
+
+            // Some blobs get pushed as packages. This is an artifact of a one-off issue in core-sdk
+            // see https://github.com/dotnet/arcade/issues/4608 for an overall fix of this.
+            // For now, if we get here, ensure that we ask for the package from the right location by
+            // stripping off the leading path elements.
+            if (!_options.NoWorkarounds)
+            {
+                assetName = Path.GetFileName(assetName);
+            }
+
+            string packageContentUrl = $"https://pkgs.dev.azure.com/{feedAccount}/{feedVisibility}_apis/packaging/feeds/{feedName}/nuget/packages/{assetName}/versions/{asset.Version}/content";
 
             // feedVisibility == "" means that the feed is internal.
+            AuthenticationHeaderValue authHeader = null;
             if (string.IsNullOrEmpty(feedVisibility))
             {
                 if (string.IsNullOrEmpty(_options.AzureDevOpsPat))
@@ -940,12 +987,15 @@ namespace Microsoft.DotNet.Darc.Operations
                     _options.AzureDevOpsPat = localSettings.AzureDevOpsToken;
                 }
 
-                client.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue(
+                authHeader = new AuthenticationHeaderValue(
                     "Basic",
                     Convert.ToBase64String(Encoding.ASCII.GetBytes(string.Format("{0}:{1}", "", _options.AzureDevOpsPat))));
             }
 
-            if (await DownloadFileAsync(client, packageContentUrl, fullTargetPath, errors, downloadOutput))
+            using var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+            var cancellationToken = cancellationTokenSource.Token;
+
+            if (await DownloadFileAsync(client, packageContentUrl, authHeader, fullTargetPath, errors, downloadOutput, cancellationToken))
             {
                 return new DownloadedAsset()
                 {
@@ -1022,11 +1072,11 @@ namespace Microsoft.DotNet.Darc.Operations
         }
 
         private async Task<DownloadedAsset> DownloadBlobAsync(HttpClient client,
-                                                                Asset asset,
-                                                                AssetLocation assetLocation,
-                                                                string subPath,
-                                                                List<string> errors,
-                                                                StringBuilder downloadOutput)
+                                                            Asset asset,
+                                                            AssetLocation assetLocation,
+                                                            string subPath,
+                                                            List<string> errors,
+                                                            StringBuilder downloadOutput)
         {
             // Normalize the asset name.  Sometimes the upload to the BAR will have
             // "assets/" prepended to it and sometimes not (depending on the Maestro tasks version).
@@ -1057,24 +1107,31 @@ namespace Microsoft.DotNet.Darc.Operations
                 string finalBaseUri = assetLocation.Location.Substring(0, assetLocation.Location.Length - "index.json".Length);
                 string finalUri1 = $"{finalBaseUri}{asset.Name}";
                 string finalUri2 = $"{finalBaseUri}assets/{asset.Name}";
-                if (await DownloadFileAsync(client, finalUri1, fullTargetPath, errors, downloadOutput))
+
+                using var cancellationTokenSource1 = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+                using var cancellationTokenSource2 = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+
+                if (await DownloadFileAsync(client, finalUri1, null, fullTargetPath, errors, downloadOutput, cancellationTokenSource1.Token))
                 {
                     downloadedAsset.Successful = true;
                     downloadedAsset.SourceLocation = finalUri1;
                     return downloadedAsset;
                 }
-                if (await DownloadFileAsync(client, finalUri2, fullTargetPath, errors, downloadOutput))
+                if (await DownloadFileAsync(client, finalUri2, null, fullTargetPath, errors, downloadOutput, cancellationTokenSource2.Token))
                 {
                     downloadedAsset.Successful = true;
                     downloadedAsset.SourceLocation = finalUri2;
                     return downloadedAsset;
                 }
+
                 // Could be under assets/assets/ in some recent builds due to a bug in the release
                 // pipeline.
                 if (!_options.NoWorkarounds)
                 {
                     string finalUri3 = $"{finalBaseUri}assets/assets/{asset.Name}";
-                    if (await DownloadFileAsync(client, finalUri3, fullTargetPath, errors, downloadOutput))
+                    using var cancellationTokenSource3 = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+
+                    if (await DownloadFileAsync(client, finalUri3, null, fullTargetPath, errors, downloadOutput, cancellationTokenSource3.Token))
                     {
                         downloadedAsset.Successful = true;
                         downloadedAsset.SourceLocation = finalUri3;
@@ -1083,7 +1140,9 @@ namespace Microsoft.DotNet.Darc.Operations
 
                     // Could also not be under /assets, so strip that from the url
                     string finalUri4 = finalUri1.Replace("assets/", "", StringComparison.OrdinalIgnoreCase);
-                    if (await DownloadFileAsync(client, finalUri4, fullTargetPath, errors, downloadOutput))
+                    using var cancellationTokenSource4 = new CancellationTokenSource(TimeSpan.FromSeconds(_options.AssetDownloadTimeoutInSeconds));
+
+                    if (await DownloadFileAsync(client, finalUri4, null, fullTargetPath, errors, downloadOutput, cancellationTokenSource4.Token))
                     {
                         downloadedAsset.Successful = true;
                         downloadedAsset.SourceLocation = finalUri4;
@@ -1152,15 +1211,23 @@ namespace Microsoft.DotNet.Darc.Operations
         }
 
         /// <summary>
-        ///     Download a single file and write it to targetFile. For now we just support
-        ///     unauthenticated blob storage. Later there could be a version that uses the storage
-        ///     SDK.
+        ///     Download a single file and write it to targetFile. If suffixes were passed in
+        ///     to darc, use those if the unauthenticated download fails.
         /// </summary>
         /// <param name="client">Http client</param>
         /// <param name="sourceUri">Source uri</param>
         /// <param name="targetFile">Target file path. Directories are created.</param>
-        /// <returns>Error message if the </returns>
-        private async Task<bool> DownloadFileAsync(HttpClient client, string sourceUri, string targetFile, List<string> errors, StringBuilder downloadOutput)
+        /// <param name="authHeader">Optional authentication header if necessary</param>
+        /// <param name="downloadOutput">Console output for the download.</param>
+        /// <param name="errors">List of errors. Append error messages to this list if there are failures.</param>
+        /// <returns>True if the download succeeded, false otherwise.</returns>
+        private async Task<bool> DownloadFileAsync(HttpClient client, 
+            string sourceUri, 
+            AuthenticationHeaderValue authHeader, 
+            string targetFile, 
+            List<string> errors, 
+            StringBuilder downloadOutput,
+            CancellationToken cancellationToken)
         {
             if (_options.DryRun)
             {
@@ -1168,6 +1235,48 @@ namespace Microsoft.DotNet.Darc.Operations
                 return true;
             }
 
+            if (_options.SkipExisting && File.Exists(targetFile))
+            {
+                return true;
+            }
+
+            if (await DownloadFileImplAsync(client, sourceUri, authHeader, targetFile, errors, downloadOutput, cancellationToken))
+            {
+                return true;
+            }
+            else
+            {
+                // Append and attempt to use the suffixes that were passed in to download from the uri
+                foreach (string sasSuffix in _options.SASSuffixes)
+                {
+                    if (await DownloadFileImplAsync(client, $"{sourceUri}{sasSuffix}", authHeader, targetFile, errors, downloadOutput, cancellationToken))
+                    {
+                        return true;
+                    }
+                }
+            }
+
+            return false;
+        }
+
+        /// <summary>
+        ///     Download a single file and write it to targetFile.
+        /// </summary>
+        /// <param name="client">Http client</param>
+        /// <param name="sourceUri">Source uri</param>
+        /// <param name="authHeader">Optional authentication header if necessary</param>
+        /// <param name="targetFile">Target file path. Directories are created.</param>
+        /// <param name="errors">List of errors. Append error messages to this list if there are failures.</param>
+        /// <param name="downloadOutput">Console output for the download.</param>
+        /// <returns>True if the download succeeded, false otherwise.</returns>
+        private async Task<bool> DownloadFileImplAsync(HttpClient client, 
+            string sourceUri, 
+            AuthenticationHeaderValue authHeader, 
+            string targetFile, 
+            List<string> errors, 
+            StringBuilder downloadOutput,
+            CancellationToken cancellationToken)
+        {
             // Use a temporary in progress file name so we don't end up with corrupted
             // half downloaded files.
             string temporaryFileName = $"{targetFile}.inProgress";
@@ -1176,26 +1285,43 @@ namespace Microsoft.DotNet.Darc.Operations
                 File.Delete(temporaryFileName);
             }
 
+            HttpRequestMessage requestMessage = null;
+
             try
             {
-                if (_options.SkipExisting && File.Exists(targetFile))
-                {
-                    return true;
-                }
-
                 string directory = Path.GetDirectoryName(targetFile);
                 Directory.CreateDirectory(directory);
+
+                if (authHeader != null)
+                {
+                    requestMessage = new HttpRequestMessage(HttpMethod.Get, sourceUri)
+                    {
+                        Headers =
+                        {
+                            { HttpRequestHeader.Authorization.ToString(), authHeader.ToString() }
+                        }
+                    };
+                }
+                else
+                {
+                    requestMessage = new HttpRequestMessage(HttpMethod.Get, sourceUri);
+                }
 
                 // Ensure the parent target directory has been created.
                 using (FileStream outStream = new FileStream(temporaryFileName,
                                                       _options.Overwrite ? FileMode.Create : FileMode.CreateNew,
                                                       FileAccess.Write))
                 {
-                    using (var inStream = await client.GetStreamAsync(sourceUri))
+                    using (var response = await client.SendAsync(requestMessage))
                     {
-                        downloadOutput.Append($"  {sourceUri} => {targetFile}...");
-                        await inStream.CopyToAsync(outStream);
-                        downloadOutput.AppendLine("Done");
+                        response.EnsureSuccessStatusCode();
+
+                        using (var inStream = await response.Content.ReadAsStreamAsync())
+                        {
+                            downloadOutput.Append($"  {sourceUri} => {targetFile}...");
+                            await inStream.CopyToAsync(outStream, cancellationToken);
+                            downloadOutput.AppendLine("Done");
+                        }
                     }
                 }
 
@@ -1212,14 +1338,26 @@ namespace Microsoft.DotNet.Darc.Operations
             {
                 // Ensure we delete the file in this case, otherwise an attempt to download
                 // from a separate location will fail.
-                File.Delete(targetFile);
+                if (File.Exists(targetFile))
+                {
+                    File.Delete(targetFile);
+                }
                 errors.Add($"Failed to download {sourceUri}: {e.Message}");
+            }
+            catch (OperationCanceledException e)
+            {
+                errors.Add($"The download operation was cancelled: {e.Message}");
             }
             finally
             {
                 if (File.Exists(temporaryFileName))
                 {
                     File.Delete(temporaryFileName);
+                }
+
+                if (requestMessage != null)
+                {
+                    requestMessage.Dispose();
                 }
             }
             return false;
